refactor: move to expo router for navigation
