+++ .gitignore
@@ -1,17 +1,33 @@
-# OSX
-#
+# macOS
 .DS_Store
 
-# XDE
-.expo/
-
-# VSCode
+# Editor / IDE
 .vscode/
 jsconfig.json
+.idea
 
-# Xcode
-#
+# JavaScript / Node
+node_modules/
+npm-debug.log
+npm-debug.*
+yarn-debug.*
+yarn-error.*
+
+# Expo / React Native
+.expo/
+dist/
+web-build/
+expo-env.d.ts
+.metro-health-check*
+
+# Turborepo / build artifacts
+.turbo/
+lib/
 build/
+coverage/
+*.tsbuildinfo
+
+# Xcode / iOS
 *.pbxuser
 !default.pbxuser
 *.mode1v3
@@ -29,54 +45,46 @@ DerivedData
 *.xcuserstate
 project.xcworkspace
 **/.xcode.env.local
+example/ios/Pods
+ios/generated
 
-# Android/IJ
-#
+# Android
 .classpath
 .cxx
 .gradle
-.idea
 .project
 .settings
 local.properties
 android.iml
+android/app/libs
+android/keystores/debug.keystore
+android/generated
 
-# Cocoapods
-#
-example/ios/Pods
+# Native generated folders (Expo prebuild)
+ios/
+android/
+
+# Buck
+buck-out/
+\.buckd/
 
 # Ruby
 example/vendor/
 
-# node.js
-#
-node_modules/
-npm-debug.log
+# Kotlin
+.kotlin/
 
-# BUCK
-buck-out/
-\.buckd/
-android/app/libs
-android/keystores/debug.keystore
+# Certificates / keys
+*.orig.*
+*.jks
+*.p8
+*.p12
+*.key
+*.pem
+*.mobileprovision
 
-# Expo
-.expo/
-
-# Turborepo
-.turbo/
-
-# generated by bob
-lib/
-
-# React Native Codegen
-ios/generated
-android/generated
-
-# React Native Nitro Modules
-nitrogen/
-
-# Misc
-coverage/
+# Local env files
+.env*.local
 
 # graphify
 graphify-out/cost.json
