+++ example/package.json
@@ -3,7 +3,7 @@
   "description": "Example app for react-native-circle-layout",
   "version": "1.0.0",
   "private": true,
-  "main": "index.js",
+  "main": "expo-router/entry",
   "scripts": {
     "android": "expo start --android",
     "ios": "expo start --ios",
@@ -11,20 +11,25 @@
     "start": "expo start"
   },
   "dependencies": {
-    "@expo/metro-runtime": "~56.0.14",
+    "@expo/metro-runtime": "~56.0.15",
     "@expo/vector-icons": "^15.1.1",
     "@react-native-community/slider": "^5.2.0",
-    "@react-navigation/native": "^6.1.18",
-    "@react-navigation/native-stack": "^6.11.0",
     "@shopify/restyle": "^2.4.5",
-    "expo": "~56.0.9",
+    "expo": "~56.0.11",
+    "expo-router": "~56.2.10",
+    "expo-splash-screen": "~56.0.10",
     "expo-status-bar": "~56.0.4",
+    "expo-system-ui": "~56.0.5",
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "react-native": "0.85.3",
+    "react-native-gesture-handler": "~2.31.1",
+    "react-native-reanimated": "4.3.1",
     "react-native-safe-area-context": "~5.7.0",
     "react-native-screens": "4.25.2",
-    "react-native-web": "~0.21.2"
+    "react-native-svg": "15.15.4",
+    "react-native-web": "~0.21.0",
+    "react-native-worklets": "0.8.3"
   },
   "devDependencies": {
     "eas-cli": "^16.31.0",
