--- example/src/navigation/constants.ts
@@ -1,6 +1,0 @@
-export enum RootScreens {
-  PLAYGROUND = 'Playground',
-  HOME = 'Home',
-  RADIAL_MENU = 'RadialMenu',
-  NESTED_RING_MENU = 'NestedRingMenu',
-}
