--- example/src/navigation/index.ts
@@ -1,2 +1,0 @@
-export { RootScreens } from './constants';
-export type { RootStackParamList, RootStackScreenProps } from './types';
