--- example/src/screens/index.ts
@@ -1,4 +1,0 @@
-export { default as Home } from './Home/Home';
-export { default as Playground } from './Playground/Playground';
-export { default as RadialMenu } from './RadialMenu/RadialMenu';
-export { default as NestedRingMenu } from './NestedRingMenu/NestedRingMenu';
